separate incremental reader in a file
